langchain-rag-types/pyproject.toml
[project]
name = "langchain-rag-types"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "networkx>=3.6.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
